+++ src/components/Main/Company/Table/RowDetail/Parents/Parents.jsx
@@ -1,23 +1,19 @@
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { PiTreeStructureFill } from "react-icons/pi";
 import { $Parents, $Message, $Title } from "./Parents.styles.jsx";
 import { $Line } from "../Childs/Childs.styles.jsx";
 import Select from "./Select/Select.jsx";
+import { CompanyContext } from "../../../../../../store/company-context.jsx";
 
 const ENTITY_HAVE_NO_PARENTS = "The entity does not have a parent.";
 
 export default function Parents({ entityName, entity }) {
-  const [parents, setParents] = useState([]);
+  // const [parentsObj, setParentsObj] = useState({});
+  const { parents } = useContext(CompanyContext);
 
-  useEffect(() => {
-    let acc = [];
-    for (const field in entity) {
-      if (typeof entity[field] === "object" && entity[field] !== null) {
-        acc = [...acc, { name: field, value: entity[field] }];
-      }
-    }
-    setParents((_prevValue) => acc);
-  }, [entity]);
+  // useEffect(() => {
+  //   setParentsObj((_prevValue) => parents);
+  // }, [entityName, entity]);
 
   function capitalizeFirstLetter(string) {
     return string.charAt(0).toUpperCase() + string.slice(1);
@@ -27,23 +23,37 @@ export default function Parents({ entityName, entity }) {
     <>
       <$Line />
       <$Parents>
-        {parents.length === 0 ? (
+        {Object.keys(parents).length === 0 ? (
           <$Message>{ENTITY_HAVE_NO_PARENTS}</$Message>
         ) : (
-          parents.map((parent) => (
+          Object.keys(parents).map((field) => (
             <>
               <$Title $color={entityName}>
                 <PiTreeStructureFill />
-                {capitalizeFirstLetter(parent.name)}
+                {capitalizeFirstLetter(field)}
               </$Title>
-              {parent.name === "memories" ? (
-                parent.value &&
-                parent.value.map((memory) => <>{memory.partNumber}</>)
+              {field === "memories" ? (
+                parents[field] &&
+                parents[field].map((memory) => <>{memory.partNumber}</>)
               ) : (
-                <Select parent={parent} entityName={entityName} parentName={parent.name}/>
+                <Select parent={parents[field]} entityName={entityName} parentName={field}/>
               )}
             </>
           ))
+          // parents.map((parent) => (
+          //   <>
+          //     <$Title $color={entityName}>
+          //       <PiTreeStructureFill />
+          //       {capitalizeFirstLetter(parent.name)}
+          //     </$Title>
+              // {parent.name === "memories" ? (
+              //   parent.value &&
+              //   parent.value.map((memory) => <>{memory.partNumber}</>)
+              // ) : (
+              //   <Select parent={parent} entityName={entityName} parentName={parent.name}/>
+              // )}
+          //   </>
+          // ))
         )}
       </$Parents>
     </>